Add two new checks to the event loop.  Both the checks are to help handle the situation where an online program cannot keep up with data taking.  The two checks are
1) In the online event handler, we compare the current timestamp to the timestamp
of the event.  If they differ by more than 10 sec, then issue a warning to the user.
The warning messages can be disabled (for instance, if running with a front-end on 
a separate machine without NTP).

2) Also in the online event handler: we do a check to ensure that the output ROOT file
is valid; if not then we skip the event.  This situation can happen when a run is ended, but
the buffer is not emptied (as is the case if the analyzer can't keep up).  In this case 
it is mandatory that the events be dropped; otherwise the user will try to fill histograms
that no longer exist.




SVN revision: 140