Add new virtual function SetFullOutputFileName().
This is  an alternative, more complicated way of setting the output ROOT filename.
In this case the user is given the run number and the midas file name and,
from that information, constructs the output ROOT filename themselves.

Since the method needs the midas file name to be meaningful it will only work offline.


SVN revision: 155