Adding support for THttpServer (built-in ROOT web server). With THttpServer should be able to get information on all created ROOT objects through a web server.  Using TRootanaEventLoop executables, you can start the THttpServer by using '-r' option, like
./anaDisplay.exe -r8088

This start the ROOT web server on port 8088.  You can then access information from the web server by going
to

http://neut18.triumf.ca:8088/

More details on this ROOT functionality here

https://github.com/linev/jsroot/blob/master/docs/HttpServer.md

So far I've only tested this online.

This new functionality is only available in recent versions of ROOT.
In principle the Makefiles are setup to automatically check if the
functionality is there by checking 'root-config --has-http'.  Need to
check that this works.