Improvement based on Konstantin feedback.
There needs to be a way for the user to implement begin-of-run
setup in their rootana display derived executable, while also allowing the rootana display
itself to do begin-of-run setup.  Previously the user could not
have a BeginRun method without clobbering the method that rootana display
was using

Solve this by having a separate TRootanaEventLoop:BeginRunRAD() method
that is only implemented by rootana display and then the normal
TRootanaEventLoop:BeginRun() that is implemented by the user (similarly
for end-of-run).  A bit of a clunky solution, but it works.